Updating the master branch with the dev branch

+++ Pain/src/CoreRender/Renderer/Renderer3d.cpp
@@ -1,3 +1,8 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
 
 #include "CoreRender/Renderer/Renderer3d.h"
 #include "CoreFiles/LogWrapper.h"
@@ -85,7 +90,8 @@ void Renderer3d::drawIndexed(VertexArray &vertexArray, uint32_t indexCount)
 {
   uint32_t count =
       indexCount ? indexCount : vertexArray.getIndexBuffer().getCount();
-  glDrawElements(GL_TRIANGLES, count, GL_UNSIGNED_INT, nullptr);
+  glDrawElements(GL_TRIANGLES, static_cast<int32_t>(count), GL_UNSIGNED_INT,
+                 nullptr);
   glBindTexture(GL_TEXTURE_2D, 0);
 }
